Fix invalid buffer handling from TCP socket
There was a problem with processing inbound messages from tcp socket.
The buffer retrieved from socket.receive was assumed as whole message.
But you could received only the part of message or the beginning of next
message. You should firstly read whole message size in the beginning of
message and than read exact size of message.